180217_first_JB

+++ JM/DinoWorld/Assets/Scripts/BattleField/Skills/SkillAttack.cs
@@ -6,7 +6,7 @@ public class SkillAttack : MonoBehaviour {
     public string skillName;
     public string skillDescription;
     public float skillDamage;
-    public float skillCost;
+    public int skillCost;
 
     public SkillAttack chooseSkill;
 }
